requirements.txt
# --- Core Deep Learning ---
ultralytics>=8.0.0
torch>=1.7.0
torchvision>=0.8.1

# --- Numerical / Utilities ---
numpy>=1.18.5
scikit-learn>=0.24.2
tqdm>=4.41.0
PyYAML>=5.3.1

# --- Image / Video ---
opencv-python>=4.1.2
Pillow>=7.1.2
matplotlib>=3.2.2

# --- Server ---
flask>=2.0.0
flask-cors>=3.0.10
requests>=2.25.0

# --- Dashboard ---
streamlit>=1.20.0
